% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare methods}
\usage{
compare_methods(n_runs = 50)
}
\arguments{
\item{kn}{S3 object of class 'wordle_knowledge'}
}
\value{
data.frame
}
\description{
Compare methods
}
