% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/distr_wordle.R
\name{distr_wordle}
\alias{distr_wordle}
\title{Simulate n Wordle Games}
\usage{
distr_wordle(n, knowledge, ...)
}
\arguments{
\item{n}{numeric, number of games to play}

\item{knowledge}{object of S3 class 'wordle_knowledge'}
}
\value{
numeric, vector of numbers of guesses needed (or NA if not guessed in 6 or less tries)
}
\description{
Given a knowledge object, the function simulates several Wordle games. It is a wrapper to
sim_wordle.
}
