% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/embed_wordlist.R
\name{embed_wordlist}
\alias{embed_wordlist}
\title{Transforms Character To Numeric}
\usage{
embed_wordlist(wl)
}
\arguments{
\item{wl}{character, each 5 letters long, all lowercase letters}
}
\value{
numeric
}
\description{
Takes a vector of character (each with 5 letters) and transforms it to a
numeric vector of length 26*5=130.
}
\details{
E.g. 3rd letter is m (13th letter in the alphabet), then the 26*2+13=65th position of
the answer is 1.
}
