% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/knowledge.R
\name{knowledge}
\alias{knowledge}
\title{Construct Knowledge Object}
\usage{
knowledge(lang = c("en", "de", "en_short"))
}
\arguments{
\item{lang, }{character 'en' (default) or 'de'}
}
\value{
knowledge object of S3 class 'wordle_knowledge'
}
\description{
Loads wordlist of admissable words.
}
\references{
en: https://raw.githubusercontent.com/tabatkins/wordle-list/main/words

de: https://raw.githubusercontent.com/SchulzKilian/GermanWordle/main/germandict.txt
}
