% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/learn.R
\name{learn}
\alias{learn}
\title{Incorporate Wordle's Feedback}
\usage{
learn(knowledge, guess, reply)
}
\arguments{
\item{knowledge}{object of S3 class 'wordle_knowledge'}

\item{guess}{character, the guess}

\item{reply}{character, wordle's reply (5 characters out of 'f', 't', 'p')}
}
\value{
updated knowledge object of S3 class 'wordle_knowledge'
}
\description{
Given current knowledge, guess and Wordle's reply, this function updates the
knowledge object.
}
