% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/popularity.R
\name{popularity}
\alias{popularity}
\title{Search Bing for Words}
\usage{
popularity(fits)
}
\arguments{
\item{fits}{character, words to lookup}
}
\value{
numeric, vector of numbers of search results
}
\description{
This functions searches bing for the provided words and returns the number of search
results for each word.
}
\references{
https://github.com/KewKalustian/wordle_cracker/blob/master/script.R
}
