% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/reply.R
\name{reply}
\alias{reply}
\title{Emulates Wordle's Feedback}
\usage{
reply(guess, ans)
}
\arguments{
\item{guess}{character, the guess to check}

\item{chosen}{character, the correct answer}
}
\value{
character
}
\description{
Given guess and actual answer, this function computes a character string of length 5.
't' corresponds to green, 'f' to grey, and 'p' to yellow.
}
\references{
https://stackoverflow.com/a/71326031/216064
}
