% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/reply_matrix.R
\name{reply_matrix}
\alias{reply_matrix}
\title{Build Reply Matrix}
\usage{
reply_matrix(kn, ncl = 1)
}
\arguments{
\item{kn}{S3 object of class 'wordle_knowledge'}

\item{ncl}{numeric, number of cores. Default 1, greater values start parallel computing.}
}
\value{
matrix
}
\description{
Build Reply Matrix
}
