% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/sim_wordle.R
\name{sim_wordle}
\alias{sim_wordle}
\title{Simulate one Wordle Game}
\usage{
sim_wordle(knowledge, ...)
}
\arguments{
\item{knowledge}{object of S3 class 'wordle_knowledge'}

\item{verbose}{logical, print out diagnostic messages (default TRUE)}
}
\value{
numeric, number of guesses needed (or NA if not guessed in 6 or less tries)
}
\description{
Given a knowledge object, the function simulates one Wordle game.
}
