% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/suggest_guess.R
\name{suggest_guess}
\alias{suggest_guess}
\title{Suggest Guess}
\usage{
suggest_guess(
  knowledge,
  num_guess,
  n = 1,
  fitting_only = NA,
  method = c("prob", "full_entropy", "contrasts", "reply_entropy"),
  sample_size = 500,
  with_scores = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{knowledge}{object of S3 class 'wordle_knowledge'}

\item{num_guess}{numeric, 1..6}

\item{fitting_only}{logical, use only fitting words as candidates (TRUE) or 
 use single-letter-only words (FALSE) or 
decide by num_guess (NA, default)}

\item{method}{character, either one of "reply_entropy" (default), "contrasts",
"full_entropy", or "prob"}

\item{sample_size}{numeric (default=30), maximum number of candidates to consider}

\item{with_scores}{logical, return words only (FALSE, default), or include scores}

\item{verbose}{logical, print out diagnostic messages (default TRUE)}
}
\value{
updated knowledge object of S3 class 'wordle_knowledge'
}
\description{
Given current knowledge (including number of guesses and previous guesses/replies),
suggest some words for the next guess.
}
\details{
Complexity of computation grows quadratically with the sample_size parameter.
}
